requirements.txt
mysql.connector
datetime
discord
discord.py-Console